geplugin: create input listener class
